Add a test for `extensionConfiguration/didUpdate` (#8093)
A follow up to https://github.com/sourcegraph/cody/pull/8079#top

The problem is that we want to control both the client in the agent in
the tests.

Initially, I tried to use the embedded agent in TestClient. Sadly, there
are more complex dependencies that I could not pass (related to
`ExtensionActivate`). Eventually, the approach is different. I added an
endpoint to handle a custom instruction from a test case.

## Test plan
- green CI

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->