Improve autoedit onboarding message (#8137)
It turned out that the configuration update in the agent does not
reflect the completion provider change for JB client.

As a quick workaround, let's instruct users to restart their IDEs.

Ideally, the config change is reflected immediately. It works for VSC. 

## Test plan
Tested manually. Sadly, there is no convenient way to reset local
storage values (we use them to store that the enrolment happened).

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->