Introduce `testing/resetStroage` endpoint and add a related action to JB (#8144)
There are some functionalities that use the local storage. We want to be
able to test them. One of them is auto-edit enrolment that uses storage
to mark a user as enrolled. Thanks to this change, a user will be able
to reset the storage from the client and retry the features.

## Test plan
1. cmd + shift + A
2. Search for: "Cody: Reset Cody Storage"
3. Execute



<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->