Update gradle toolchains resolver (#8186)
## Changes

This should fix the problem visible on other builds on the CI:

```
* What went wrong:
Could not determine the dependencies of task ':compileKotlin'.
> Cannot find a Java installation on your machine matching this tasks requirements: {languageVersion=17, vendor=JetBrains, implementation=vendor-specific} for LINUX on x86_64.
   > No matching toolchain could be found in the locally installed toolchains or the configured toolchain download repositories.
```

Looks like resolver we are using is too old, but I'm not sure why it
started to fail just now (I guess it must be some change on the
toolchains repositories side).

## Test plan

Green CI

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->