chore(ci): resolve issue with artifacts naming  (#8195)
When a failure occurs in the e2e tests, we upload some of the artifacts
for access. However, the naming of this artifact isn't unique and when
an artifact with the same name exists, it throws an error.

This PR makes the artifact name unique and overwrites it in the event
that running the workflow multiple times returns an error

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->
* CI

---------

Co-authored-by: David Ichim <ichim.david@gmail.com>
(cherry picked from commit f4280dcd00096f13b9abc7459d0085dd2ef3c51b)